fix(encoding): restore Shift_JIS/CP932 encoding for 19 files
Restored files from commit c8f6695 (before df77532) to fix UTF-8 encoding
corruption caused by Edit tool usage.

The Edit tool converted Shift_JIS bytes to UTF-8 without BOM starting from
commit df77532, causing mojibake (文字化け) when opened in VS2022.

Evidence: Bytes changed from '91 ce 89 9e' (Shift_JIS) to 'ef bf bd ce 89
ef bf bd' (UTF-8 replacement characters).

Files restored:
- SRC/ConfigFile/IniFile.cpp
- SRC/MeaSys/Xmp.cpp
- SRC/MotTsk/WIN.CPP
- SRC/NEXIO/NEXIO.CPP
- SRC/NanoSpec/* (15 files)
- SRC/Stage/STAGE.CPP

Warning fixes (C4018, C4244) will be reapplied using binary-safe methods
that preserve Shift_JIS encoding.

+++ SRC/MeaSys/Xmp.cpp
@@ -8,7 +8,7 @@
 #include "Mojiretsu.h"
 #include "Xmp.h"
 
-// XMPï¿½p
+// XMP—p
 #define DDE_SERVICE 			 		_T("jaw_adap")
 #define DDE_TOPIC				 		_T("dde")
 #define DDE_ITEM				 		_T("results")
@@ -97,7 +97,7 @@ BOOL CXmp::RunXmp()
 {
 	TRACE(_T("CXmp::RunXmp()\n"));
 
-	// XMP ï¿½Nï¿½ï¿½ï¿½iAdap ï¿½Nï¿½ï¿½ï¿½j
+	// XMP ‹N“®iAdap ‹N“®j
 	STARTUPINFO si;
 	PROCESS_INFORMATION pi;
 	char szAdapExePath[MAX_PATH];
@@ -117,30 +117,30 @@ BOOL CXmp::RunXmp()
 	else{
 		strcpy(szAdapExePath, ADAP_EXE_PATH);
 	}
-/* added 2014.05.22 hmenjo SC_adap timeout ï¿½Îï¿½ ---------- { ---------- */
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- { ---------- */
+/* added 2014.05.22 hmenjo SC_adap timeout ‘Îô ---------- { ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- { ---------- */
 //	BOOL bSC_Adap = FALSE;
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ----------              */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ----------              */
 	BOOL m_bSC_Adap = FALSE;
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- } ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- } ---------- */
 	char szLowerAdapExePath[MAX_PATH];
 	memset(szLowerAdapExePath, 0, sizeof(szLowerAdapExePath));
-	for(int i = 0; i < static_cast<int>(strlen(srXmp.szAdapExePath)); i++){
+	for(int i = 0; i < strlen(srXmp.szAdapExePath); i++){
 		szLowerAdapExePath[i] = tolower(srXmp.szAdapExePath[i]);
 	}
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- { ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- { ---------- */
 //	bSC_Adap = (strstr(szLowerAdapExePath, "sc_adap.exe") != NULL);
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ----------              */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ----------              */
 	m_bSC_Adap = (strstr(szLowerAdapExePath, "sc_adap.exe") != NULL);
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- } ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- } ---------- */
 
 	DWORD l_dwSC2cnt = 5;	/* default	*/
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- { ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- { ---------- */
 //	if (FALSE != bSC_Adap) {
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ----------              */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ----------              */
 	if (FALSE != m_bSC_Adap) {
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- } ---------- */
-		/* ï¿½Xï¿½gï¿½ï¿½ï¿½eï¿½W(ï¿½fï¿½Bï¿½ï¿½ï¿½Nï¿½gï¿½ï¿½)ï¿½ï¿½ï¿½Jï¿½Eï¿½ï¿½ï¿½g	*/
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- } ---------- */
+		/* ƒXƒgƒ‰ƒeƒW(ƒfƒBƒŒƒNƒgƒŠ)‚ğƒJƒEƒ“ƒg	*/
 		CString l_strStrategyListPath = srXmp.szStrategyListPath;
 		CString l_strStrategyDir;
 		if (-1 != l_strStrategyListPath.ReverseFind('\\')) {
@@ -168,7 +168,7 @@ BOOL CXmp::RunXmp()
 					(_T('S') != l_tszFileTitle[1]) && (_T('s') != l_tszFileTitle[1])) {
 					continue;
 				}
-				for (i = 2; i < static_cast<int>(_tcslen(l_tszFileTitle)); i++) {
+				for (i = 2; i < _tcslen(l_tszFileTitle); i++) {
 					if ((l_tszFileTitle[i] < _T('0')) || (_T('9') < l_tszFileTitle[i])) {
 						continue;
 					}
@@ -180,7 +180,7 @@ BOOL CXmp::RunXmp()
 			}
 		}
 	}
-/* added 2014.05.22 hmenjo SC_adap timeout ï¿½Îï¿½ ---------- } ---------- */
+/* added 2014.05.22 hmenjo SC_adap timeout ‘Îô ---------- } ---------- */
 
 	// Start the child process.
 	if ( !::CreateProcess(NULL, szAdapExePath, NULL, NULL, FALSE, 0, NULL, NULL, &si, &pi) )
@@ -197,18 +197,18 @@ BOOL CXmp::RunXmp()
 	CTimer timer;
 
 // 2014.03.08 bagus SCOUT modified -->
-// 	// Wvase32ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½Nï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½`ï¿½Fï¿½bï¿½N
+// 	// Wvase32‚ª³í‹N“®‚µ‚½‚©ƒ`ƒFƒbƒN
 // 	timer.Restart(10);
 // 	while ( (hWnd = ::FindWindow(NULL, srXmp.szWvaseWindowTitle)) == NULL )	{
 // 		if ( timer.IsTimeout() ) {
-// 			// WVASE32ï¿½ï¿½ï¿½ï¿½ï¿½^ï¿½Cï¿½ï¿½ï¿½Aï¿½Eï¿½gï¿½Gï¿½ï¿½ï¿½[
+// 			// WVASE32‰“šƒ^ƒCƒ€ƒAƒEƒgƒGƒ‰[
 // 			::MessageBox(NULL, pMojiretsu->LoadString(IDS_STRING22), g_szMsgBoxCaption, MB_OK | MB_ICONSTOP);
 // 			return FALSE;
 // 		}
 // 		::Sleep(10);
 // 	}
 
-/* deleted 2014.05.22 hmenjo SC_adap timeout ï¿½Îï¿½ ---------- { ---------- */
+/* deleted 2014.05.22 hmenjo SC_adap timeout ‘Îô ---------- { ---------- */
 //	BOOL bSC_Adap = FALSE;
 //
 //	char szLowerAdapExePath[MAX_PATH];
@@ -217,23 +217,23 @@ BOOL CXmp::RunXmp()
 //		szLowerAdapExePath[i] = tolower(srXmp.szAdapExePath[i]);
 //	}
 //	bSC_Adap = (strstr(szLowerAdapExePath, "sc_adap.exe") != NULL);
-/* deleted 2014.05.22 hmenjo SC_adap timeout ï¿½Îï¿½ ---------- } ---------- */
+/* deleted 2014.05.22 hmenjo SC_adap timeout ‘Îô ---------- } ---------- */
 
-	// Wvase32ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½Nï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½`ï¿½Fï¿½bï¿½N
+	// Wvase32‚ª³í‹N“®‚µ‚½‚©ƒ`ƒFƒbƒN
 	timer.Restart(60);
 
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- { ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- { ---------- */
 //	if(bSC_Adap){
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ----------              */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ----------              */
 	if (FALSE != m_bSC_Adap) {
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- } ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- } ---------- */
 		g_bScoutAppExist = FALSE;
 		strcpy(g_szWindowTitle, srXmp.szWvaseWindowTitle);
 		while(TRUE){
 			EnumWindows(EnumWindowsProc, 0);
 			if(!g_bScoutAppExist){
 				if ( timer.IsTimeout() ) {
-					// SCOUTï¿½ï¿½ï¿½ï¿½ï¿½^ï¿½Cï¿½ï¿½ï¿½Aï¿½Eï¿½gï¿½Gï¿½ï¿½ï¿½[
+					// SCOUT‰“šƒ^ƒCƒ€ƒAƒEƒgƒGƒ‰[
 					::MessageBox(NULL, pMojiretsu->LoadString(IDS_STRING24), g_szMsgBoxCaption, MB_OK | MB_ICONSTOP);
 					return FALSE;
 				}
@@ -247,7 +247,7 @@ BOOL CXmp::RunXmp()
 	else{
 		while ( (hWnd = ::FindWindow(NULL, srXmp.szWvaseWindowTitle)) == NULL )	{
 			if ( timer.IsTimeout() ) {
-				// WVASE32ï¿½ï¿½ï¿½ï¿½ï¿½^ï¿½Cï¿½ï¿½ï¿½Aï¿½Eï¿½gï¿½Gï¿½ï¿½ï¿½[
+				// WVASE32‰“šƒ^ƒCƒ€ƒAƒEƒgƒGƒ‰[
 				::MessageBox(NULL, pMojiretsu->LoadString(IDS_STRING22), g_szMsgBoxCaption, MB_OK | MB_ICONSTOP);
 				return FALSE;
 			}
@@ -258,53 +258,53 @@ BOOL CXmp::RunXmp()
 
 // 2014.03.08 bagus SCOUT modified -->
 //	::ShowWindow(hWnd, SW_HIDE);
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- { ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- { ---------- */
 //	if(!bSC_Adap){
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ----------              */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ----------              */
 	if (FALSE == m_bSC_Adap) {
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- } ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- } ---------- */
 		::ShowWindow(hWnd, SW_HIDE);
 	}
 // 2014.03.08 bagus SCOUT modified <--
 
-	// XMP DDE ï¿½ÊMï¿½Jï¿½n
+	// XMP DDE ’ÊMŠJn
 	CDdeClient xmpDde;
 	xmpDde.Init();
-/* modified 2014.05.22 hmenjo SC_adap timeout ï¿½Îï¿½ ---------- { ---------- */
+/* modified 2014.05.22 hmenjo SC_adap timeout ‘Îô ---------- { ---------- */
 //	timer.Restart(10);
-/* modified 2014.05.22 hmenjo SC_adap timeout ï¿½Îï¿½ ----------              */
+/* modified 2014.05.22 hmenjo SC_adap timeout ‘Îô ----------              */
 	double l_dAdapTimeout = 10.0;
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- { ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- { ---------- */
 //	if (FALSE != bSC_Adap) {
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ----------              */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ----------              */
 	if (FALSE != m_bSC_Adap) {
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- } ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- } ---------- */
 		double l_dTO_ScoutStart = 30.0;		/* [s]	*/
 		double l_dTO_ReadSC2 = 12.0;		/* [s]	*/
 		l_dAdapTimeout = l_dTO_ScoutStart + (l_dTO_ReadSC2 * l_dwSC2cnt);
 	}
 	timer.Restart(l_dAdapTimeout);
-/* modified 2014.05.22 hmenjo SC_adap timeout ï¿½Îï¿½ ---------- } ---------- */
+/* modified 2014.05.22 hmenjo SC_adap timeout ‘Îô ---------- } ---------- */
 	while ( !xmpDde.Connect(DDE_SERVICE, DDE_TOPIC) ) {
 		if ( timer.IsTimeout() ) {
-			// DDE Communication ï¿½^ï¿½Cï¿½ï¿½ï¿½Aï¿½Eï¿½gï¿½Gï¿½ï¿½ï¿½[
+			// DDE Communication ƒ^ƒCƒ€ƒAƒEƒgƒGƒ‰[
 			::MessageBox(NULL, pMojiretsu->LoadString(IDS_STRING23), g_szMsgBoxCaption, MB_OK | MB_ICONSTOP);
 			return FALSE;
 		}
 		else
 			::Sleep(10);
 	};
-	xmpDde.Execute(_T("WindowState(1)"));		// XMP unvisibleï¿½É‚ï¿½ï¿½ï¿½
+	xmpDde.Execute(_T("WindowState(1)"));		// XMP unvisible‚É‚·‚é
 	xmpDde.Disconnect();
 	xmpDde.Uninit();
 
 // 2014.03.08 bagus SCOUT modified -->
 //	::ShowWindow(hWnd, SW_HIDE);
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- { ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- { ---------- */
 //	if(!bSC_Adap){
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ----------              */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ----------              */
 	if (FALSE == m_bSC_Adap) {
-/* modified 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- } ---------- */
+/* modified 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- } ---------- */
 		::ShowWindow(hWnd, SW_HIDE);
 	}
 // 2014.03.08 bagus SCOUT modified <--
@@ -318,7 +318,7 @@ void CXmp::ShutDownXmp()
 {
 	TRACE(_T("CXmp::ShutDownXmp()\n"));
 
-	// XMP DDE ï¿½ÊMï¿½Iï¿½ï¿½
+	// XMP DDE ’ÊMI—¹
 	CDdeClient xmpDde;
 	xmpDde.Init();
 	xmpDde.Connect(DDE_SERVICE, DDE_TOPIC);
@@ -338,21 +338,21 @@ BOOL CXmp::DoXmp(LPTSTR pszXmpDdeText, LPCTSTR pszCmd, CDdeClient* pXmpDde)
 
 	pXmpDde->Execute(pszCmd);
 
-	// DDEï¿½ÊMï¿½ï¿½ADAPï¿½ï¿½ï¿½çŒ‹ï¿½Êæ“¾
+	// DDE’ÊM‚ÅADAP‚©‚çŒ‹‰Êæ“¾
 	if ( !pXmpDde->DataCopy(szBuff, DDE_ITEM) )	{
-		// DDEï¿½ÊMï¿½Gï¿½ï¿½ï¿½[
+		// DDE’ÊMƒGƒ‰[
 		return FALSE;
 	}
-	// WVASE32ï¿½vï¿½Zï¿½ï¿½ï¿½Ì‚ï¿½ï¿½ßŒï¿½ï¿½Ê‘Ò‚ï¿½ï¿½BADAPï¿½ÍAWVASE32.EXEï¿½ï¿½FITï¿½lï¿½ï¿½ï¿½æ“¾ï¿½ï¿½ï¿½ï¿½Ü‚Å‚ÌŠÔA""ï¿½ï¿½Ô‹pï¿½ï¿½ï¿½Ä‚ï¿½ï¿½ï¿½
+	// WVASE32ŒvZ’†‚Ì‚½‚ßŒ‹‰Ê‘Ò‚¿BADAP‚ÍAWVASE32.EXE‚ÌFIT’l‚ğæ“¾‚·‚é‚Ü‚Å‚ÌŠÔA""‚ğ•Ô‹p‚µ‚Ä‚­‚é
 	timer.Restart(ADAPTIMEOUT);
 	while ( _tcscmp(szBuff, _T("")) == 0 ) {
 		if ( timer.IsTimeout() ) {
-			// WVASE32ï¿½ï¿½ï¿½ï¿½ï¿½^ï¿½Cï¿½ï¿½ï¿½Aï¿½Eï¿½gï¿½Gï¿½ï¿½ï¿½[
+			// WVASE32‰“šƒ^ƒCƒ€ƒAƒEƒgƒGƒ‰[
 			return FALSE;
 		}
 		::Sleep(10);
 		if ( !pXmpDde->DataCopy(szBuff, DDE_ITEM) )	{
-			// DDEï¿½ÊMï¿½Gï¿½ï¿½ï¿½[
+			// DDE’ÊMƒGƒ‰[
 			return FALSE;
 		}
 	}
@@ -366,7 +366,7 @@ BOOL CXmp::DoXmp(LPTSTR pszXmpDdeText, LPCTSTR pszCmd, CDdeClient* pXmpDde)
 	return TRUE;
 }
 
-/* added 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- { ---------- */
+/* added 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- { ---------- */
 BOOL CXmp::ShowXmp(int iShow)
 {
 	TRACE(_T("CXmp::ShowXmp()\n"));
@@ -379,7 +379,7 @@ BOOL CXmp::ShowXmp(int iShow)
 		case 1:
 		case 2:
 			{
-				/* xmp ï¿½É•\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½	*/
+				/* xmp ‚É•\¦/”ñ•\¦	*/
 				int l_iCmd = 1;
 				switch (iShow) {
 				case 1:		l_iCmd = 1;	break;
@@ -393,16 +393,16 @@ BOOL CXmp::ShowXmp(int iShow)
 				l_xmpDde.Execute(l_tszCmd);
 				l_xmpDde.Disconnect();
 				l_xmpDde.Uninit();
-				/* ADAP ï¿½É•\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½	*/
+				/* ADAP ‚É•\¦/”ñ•\¦	*/
 				::ShowWindow(l_hWnd, (1 == l_iCmd)? SW_HIDE : SW_SHOWNORMAL);
 			}
 			break;
 		default:
-			/* ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½È‚ï¿½	*/
+			/* ‰½‚à‚µ‚È‚¢	*/
 			break;
 		}
 	}
 
 	return l_bRc;
 }
-/* added 2014.11.25 hmenjo XMP ï¿½\ï¿½ï¿½/ï¿½ï¿½\ï¿½ï¿½ ---------- } ---------- */
+/* added 2014.11.25 hmenjo XMP •\¦/”ñ•\¦ ---------- } ---------- */
